Plugin poc for setupo init command
