+++ pkg/cmd/enabler/setup/initcmd/asdf.go
@@ -1,0 +1,1 @@
+package initcmd
